Add files via upload
This repository has the java and .xhml(web) files. Working of the project can be seen in the YouTube video . Lastly, In order to run this project as a whole we need mysql connector 4.1.48 for connecting to database and Bootfaces- OSP- 1.5.0 to make Bootstrap content work on the Front-End side.